Implement VAR for C21_WMDE_ipad_Test_02
