+++ pad/components/Banner_var.jsx
@@ -1,64 +1,47 @@
-import { Component, h, createRef } from 'preact';
+import { Component, h } from 'preact';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
 import BannerTransition from '../../shared/components/BannerTransition';
 import ProgressBar from '../../shared/components/ui/ProgressBar';
-import Footer from '../../shared/components/ui/EasySelectFooter';
+import Slider from '../../shared/components/Slider';
+import Slides from './Slides';
+import SlideState from '../../shared/slide_state';
+import ChevronLeftIcon from './ui/ChevronLeftIcon_var';
+import ChevronRightIcon from './ui/ChevronRightIcon_var';
+import debounce from '../../shared/debounce';
 import TranslationContext from '../../shared/components/TranslationContext';
-import { LocalImpressionCount } from '../../shared/local_impression_count';
-import { CampaignProjection } from '../../shared/campaign_projection';
 import FundsDistributionInfo from '../../shared/components/ui/use_of_funds/FundsDistributionInfo';
 import FundsModal from '../../shared/components/ui/use_of_funds/FundsModal';
-import Slider from '../../shared/components/Slider';
+import Footer from '../../shared/components/ui/EasySelectFooter';
 import { BannerType } from '../../shared/BannerType';
 import createDynamicCampaignText from '../create_dynamic_campaign_text';
-import Slides from './Slides';
-import ChevronLeftIcon from './ui/ChevronLeftIcon';
-import ChevronRightIcon from './ui/ChevronRightIcon';
-import SlideState from '../../shared/slide_state';
-import debounce from '../../shared/debounce';
-
-const PENDING = 0;
-const VISIBLE = 1;
-const CLOSED = 2;
 
 const SLIDESHOW_START_DELAY = 2000;
 const SLIDESHOW_SLIDE_INTERVAL = 10000;
 
+const BannerVisibilityState = Object.freeze( {
+	PENDING: Symbol( 'pending' ),
+	VISIBLE: Symbol( 'visible' ),
+	CLOSED: Symbol( 'closed' )
+} );
+
 export default class Banner extends Component {
 
 	static propTypes = {
-		bannerName: PropTypes.string,
-		campaignName: PropTypes.string,
-
-		campaignParameters: PropTypes.object,
-		campaignProjection: PropTypes.instanceOf( CampaignProjection ),
-		formatters: PropTypes.object,
-		donationForm: PropTypes.elementType,
-		translations: PropTypes.object,
-		/**
-		 *  Contains items with values and labels for SelectGroup components
-		 *  in the form, e.g. amounts, payment types, intervals, etc
-		 */
-		formItems: PropTypes.object,
-
-		/** object that holds any data needed for tracking purposes */
-		trackingData: PropTypes.object,
-		impressionCounts: PropTypes.instanceOf( LocalImpressionCount ),
-
 		/** callback when banner closes */
 		onClose: PropTypes.func,
 		/** Callback to register a displayBanner function with the BannerPresenter */
-		registerDisplayBanner: PropTypes.func.isRequired
+		registerDisplayBanner: PropTypes.func.isRequired,
+		/** callback when banner gets submitted */
+		onSubmit: PropTypes.func
 	}
-
-	ref = createRef();
 
 	constructor( props ) {
 		super( props );
 		this.state = {
-			displayState: PENDING,
+			bannerVisibilityState: BannerVisibilityState.PENDING,
 			isFundsModalVisible: false,
+			contentSize: 'auto',
 
 			// trigger for banner resize events
 			formInteractionSwitcher: false
@@ -79,23 +62,30 @@ export default class Banner extends Component {
 	componentDidMount() {
 		this.props.registerDisplayBanner(
 			() => {
-				this.setState( { displayState: VISIBLE } );
+				this.setState( { bannerVisibilityState: BannerVisibilityState.VISIBLE } );
 				this.slideInBanner();
 			}
 		);
 		this.props.registerResizeBanner( debounce( this.onPageResize.bind( this ), 200 ) );
 	}
 
-	trackBannerEvent( eventName, trackRatio ) {
+	trackBannerEvent( eventName ) {
 		this.props.trackingData.tracker.trackBannerEvent(
 			eventName,
 			this.slideState.slidesShown,
 			this.slideState.currentSlide + 1,
-			trackRatio ?? this.props.trackingData.bannerClickTrackRatio
+			this.props.trackingData.bannerClickTrackRatio
 		);
 	}
 
 	onPageResize() {
+		if ( this.state.bannerVisibilityState !== BannerVisibilityState.VISIBLE ) {
+			return;
+		}
+		this.addBannerSpace();
+	}
+
+	addBannerSpace() {
 		const bannerElement = document.querySelector( '.wmde-banner .banner-position' );
 		this.props.skinAdjuster.addSpaceInstantly( bannerElement.offsetHeight );
 	}
@@ -108,15 +98,19 @@ export default class Banner extends Component {
 	}
 
 	onFinishedTransitioning = () => {
-		this.startProgressbar();
 		this.props.onFinishedTransitioning();
+		this.startProgressbar();
 		setTimeout( this.startSliderAutoplay, SLIDESHOW_START_DELAY );
-		this.onPageResize();
+	}
+
+	onSubmit = () => {
+		this.trackBannerEvent( 'submit' );
+		this.props.onSubmit();
 	}
 
 	closeBanner = e => {
 		e.preventDefault();
-		this.setState( { displayState: CLOSED } );
+		this.setState( { bannerVisibilityState: BannerVisibilityState.CLOSED } );
 		this.props.onClose();
 	};
 
@@ -146,8 +140,8 @@ export default class Banner extends Component {
 	};
 
 	onFormInteraction = () => {
-		this.setState( { showLanguageWarning: true, formInteractionSwitcher: !this.state.formInteractionSwitcher } );
 		this.stopSliderAutoplay();
+		this.setState( { showLanguageWarning: false, formInteractionSwitcher: !this.state.formInteractionSwitcher } );
 	}
 
 	onSlideChange = ( index ) => {
@@ -162,13 +156,12 @@ export default class Banner extends Component {
 			className={ classNames(
 				'wmde-banner',
 				{
-					'wmde-banner--hidden': state.displayState === CLOSED,
-					'wmde-banner--visible': state.displayState === VISIBLE,
+					'wmde-banner--hidden': state.bannerVisibilityState === BannerVisibilityState.CLOSED,
+					'wmde-banner--visible': state.bannerVisibilityState === BannerVisibilityState.VISIBLE,
 					'wmde-banner--ctrl': props.bannerType === BannerType.CTRL,
 					'wmde-banner--var': props.bannerType === BannerType.VAR
 				}
-			) }
-			ref={this.ref}>
+			) }>
 			<BannerTransition
 				fixed={ true }
 				registerDisplayBanner={ this.registerBannerTransition }
@@ -178,41 +171,42 @@ export default class Banner extends Component {
 			>
 				<TranslationContext.Provider value={props.translations}>
 					<div className="banner__wrapper">
-						<div className="banner__content">
-							<div className="banner__infobox">
-								<div className="banner__slideshow">
-									<Slider
-										slides={ Slides( this.dynamicCampaignText ) }
-										onSlideChange={ this.onSlideChange }
-										registerAutoplay={ this.registerAutoplayCallbacks }
-										interval={ SLIDESHOW_SLIDE_INTERVAL }
-										previous={ <ChevronLeftIcon/> }
-										next={ <ChevronRightIcon/> }
-									/>
-									<ProgressBar
-										formatters={props.formatters}
-										daysLeft={campaignProjection.getRemainingDays()}
-										donationAmount={campaignProjection.getProjectedDonationSum()}
-										goalDonationSum={campaignProjection.goalDonationSum}
-										missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
-										setStartAnimation={this.registerStartProgressbar}/>
-								</div>
-							</div>
-							<DonationForm
-								formItems={props.formItems}
-								bannerName={props.bannerName}
-								campaignName={props.campaignName}
-								formatters={props.formatters}
-								impressionCounts={props.impressionCounts}
-								onFormInteraction={this.onFormInteraction}
-								customAmountPlaceholder={ props.translations[ 'custom-amount-placeholder' ] }
-								onSubmit={ props.onSubmit }
-							/>
-						</div>
 						<div className="close">
-							<a className="close__link" onClick={ this.closeBanner }>&#x2715;</a>
+							<a className="close__link" onClick={this.closeBanner}>&#x2715;</a>
+						</div>
+						<div className="banner__content">
+							<div className="banner__slideshow">
+								<Slider
+									slides={ Slides( this.dynamicCampaignText ) }
+									onSlideChange={ this.onSlideChange }
+									registerAutoplay={ this.registerAutoplayCallbacks }
+									interval={ SLIDESHOW_SLIDE_INTERVAL }
+									previous={ <ChevronLeftIcon/> }
+									next={ <ChevronRightIcon/> }
+								/>
+							</div>
+							<div className="banner__form">
+								<DonationForm
+									formItems={props.formItems}
+									bannerName={props.bannerName}
+									campaignName={props.campaignName}
+									formatters={props.formatters}
+									impressionCounts={props.impressionCounts}
+									onFormInteraction={this.onFormInteraction}
+									customAmountPlaceholder={ props.translations[ 'custom-amount-placeholder' ] }
+									onSubmit={ this.onSubmit }
+									showFundsModal={ this.toggleFundsModal }
+								/>
+							</div>
 						</div>
 						<Footer showFundsModal={ this.toggleFundsModal }/>
+						<ProgressBar
+							formatters={props.formatters}
+							daysLeft={campaignProjection.getRemainingDays()}
+							donationAmount={campaignProjection.getProjectedDonationSum()}
+							goalDonationSum={campaignProjection.goalDonationSum}
+							missingAmount={campaignProjection.getProjectedRemainingDonationSum()}
+							setStartAnimation={this.registerStartProgressbar}/>
 					</div>
 				</TranslationContext.Provider>
 			</BannerTransition>
@@ -227,5 +221,4 @@ export default class Banner extends Component {
 			</FundsModal>
 		</div>;
 	}
-
 }
