+++ pad/components/Slides.jsx
