+++ pad/components/ui/ChevronLeftIcon_var.jsx
@@ -1,0 +1,8 @@
+import { h } from 'preact';
+
+export default function ChevronLeftIcon() {
+	return <svg width="9" height="16" viewBox="0 0 9 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+		{/* eslint-disable-next-line max-len */}
+		<path d="M0.242482 7.41446C-0.0808273 7.73785 -0.0808275 8.26215 0.242482 8.58554L7.41279 15.7575C7.74401 16.0888 8.28382 16.0794 8.60343 15.7369L8.77734 15.5506C9.08027 15.2259 9.07325 14.7201 8.76143 14.4039L2.63553 8.19386C2.5295 8.08638 2.5295 7.91362 2.63553 7.80614L8.76143 1.59605C9.07325 1.27995 9.08027 0.774089 8.77734 0.449448L8.60343 0.263065C8.28382 -0.0794487 7.74401 -0.0887552 7.41279 0.242538L0.242482 7.41446Z" fill="#202122"/>
+	</svg>;
+}
