+++ pad/components/ui/ChevronRightIcon_var.jsx
@@ -1,0 +1,8 @@
+import { h } from 'preact';
+
+export default function ChevronRightIcon() {
+	return <svg width="9" height="16" viewBox="0 0 9 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+		{/* eslint-disable-next-line max-len */}
+		<path d="M8.75752 8.58554C9.08083 8.26215 9.08083 7.73785 8.75752 7.41446L1.58721 0.242538C1.25599 -0.088754 0.71618 -0.079448 0.396574 0.263065L0.222655 0.449448C-0.080274 0.77409 -0.0732549 1.27995 0.238566 1.59605L6.36447 7.80614C6.4705 7.91362 6.4705 8.08638 6.36447 8.19386L0.238567 14.4039C-0.0732528 14.7201 -0.0802727 15.2259 0.222656 15.5506L0.396575 15.7369C0.716182 16.0794 1.25599 16.0888 1.58721 15.7575L8.75752 8.58554Z" fill="#202122"/>
+	</svg>;
+}
