+++ pad/components/ui/InfoIcon.jsx
@@ -1,0 +1,8 @@
+import { h } from 'preact';
+
+export default function InfoIcon() {
+	return <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+		<circle cx="8" cy="8" r="8" fill="#2B6DA0"/>
+		<path d="M8.80003 5.73332V11.7173H10.2667V13.0667H5.96937V11.7173H7.33337V7.19999H5.8667V5.73332H8.80003ZM7.33337 2.79999H8.80003V4.26665H7.33337V2.79999Z" fill="white"/>
+	</svg>;
+}
