--- pad/components_var/TranslationContext.js
@@ -1,4 +1,0 @@
-import { createContext } from 'preact/compat';
-import Translations from '../../shared/messages/de';
-
-export default createContext( Translations );
