--- pad/components_var/ui/ChevronRightIcon.jsx
@@ -1,8 +1,0 @@
-import { h } from 'preact';
-
-export default function ChevronRightIcon() {
-	return <svg width="15" height="23" viewBox="0 0 15 23" fill="none" xmlns="http://www.w3.org/2000/svg">
-		{/* eslint-disable-next-line max-len */}
-		<path d="M3.0006 22.9996C2.2046 22.9996 1.4096 22.6846 0.818601 22.0596C-0.318399 20.8546 -0.264399 18.9566 0.939601 17.8186L7.6306 11.4996L0.939601 5.1806C-0.264399 4.0426 -0.318399 2.1446 0.818601 0.939601C1.9576 -0.264399 3.8546 -0.318399 5.0606 0.818601L14.0606 9.3186C14.6606 9.8856 15.0006 10.6746 15.0006 11.4996C15.0006 12.3246 14.6606 13.1136 14.0606 13.6806L5.0606 22.1806C4.4796 22.7286 3.7396 22.9996 3.0006 22.9996Z" fill="#3366CC"/>
-	</svg>;
-}
