--- pad/components_var/ui/EditIcon.jsx
@@ -1,12 +1,0 @@
-/* eslint-disable react/no-unknown-property,max-len */
-import { h } from 'preact';
-
-export default function EditIcon() {
-	return <svg width="15" height="18" viewBox="0 0 15 18" fill="none" xmlns="http://www.w3.org/2000/svg">
-		<mask id="path-1-inside-1" fill="white">
-			<path d="M6.22083 12.5497C6.15129 12.6326 6.05675 12.6907 5.95139 12.7152L4.3283 13.0938C4.00368 13.1696 3.69726 12.9125 3.71546 12.5796L3.80647 10.9154C3.81237 10.8074 3.85316 10.7042 3.9227 10.6213L11.8504 1.17347C12.0279 0.961928 12.3433 0.934336 12.5548 1.11184L14.0869 2.39741C14.2985 2.57491 14.326 2.89029 14.1485 3.10183L6.22083 12.5497Z"/>
-		</mask>
-		<path d="M3.80647 10.9154L0.810942 10.7516L3.80647 10.9154ZM5.2699 9.79367L3.64681 10.1723L5.00979 16.0154L6.63288 15.6368L5.2699 9.79367ZM6.71098 12.7434L6.80199 11.0792L0.810942 10.7516L0.719936 12.4158L6.71098 12.7434ZM6.22083 12.5497L14.1485 3.10183L9.55228 -0.754898L1.62456 8.69298L6.22083 12.5497ZM10.6265 3.40997L12.1586 4.69555L16.0153 0.0992789L14.4832 -1.1863L10.6265 3.40997ZM11.8504 1.17347L3.9227 10.6213L8.51897 14.4781L16.4467 5.03019L11.8504 1.17347ZM12.1586 4.69555C11.1009 3.80804 10.9629 2.23115 11.8504 1.17347L16.4467 5.03019C17.6892 3.54943 17.496 1.34179 16.0153 0.0992789L12.1586 4.69555ZM14.1485 3.10183C13.261 4.15951 11.6842 4.29748 10.6265 3.40997L14.4832 -1.1863C13.0024 -2.4288 10.7948 -2.23566 9.55228 -0.754898L14.1485 3.10183ZM6.80199 11.0792C6.77245 11.6194 6.56853 12.1353 6.22083 12.5497L1.62456 8.69298C1.13778 9.27311 0.852293 9.99546 0.810942 10.7516L6.80199 11.0792ZM3.64681 10.1723C5.2699 9.79367 6.80199 11.0792 6.71098 12.7434L0.719936 12.4158C0.592528 14.7457 2.73745 16.5455 5.00979 16.0154L3.64681 10.1723ZM6.63288 15.6368C7.37038 15.4648 8.03219 15.0582 8.51897 14.4781L3.9227 10.6213C4.2704 10.207 4.74312 9.91655 5.2699 9.79367L6.63288 15.6368Z" fill="#535353" mask="url(#path-1-inside-1)"/>
-		<path d="M1.5 16.5H13.5" stroke="#535353" stroke-width="2" stroke-linecap="square"/>
-	</svg>;
-}
