+++ pad/styles/SelectGroup.pcss
@@ -107,6 +107,9 @@
 				line-height: 2em;
 				font-size: 0.74em;
 				text-align: center;
+				outline: 0;
+				background: transparent;
+				border: 0;
 			}
 
 			&--euro-symbol {
