+++ pad/styles_var/Footer.pcss
@@ -1,26 +1,27 @@
 .wmde-banner {
 	.footer {
 		display: flex;
-		flex-wrap: wrap;
 		flex-direction: row;
 		justify-content: flex-start;
 		font-size: 12px;
-		line-height: 1.6em;
+		line-height: 20px;
 		color: $color-black;
-		text-shadow: -1px -1px 0 $color-white, 1px -1px 0 $color-white, -1px 1px 0 $color-white, 1px 1px 0 $color-white;
+		padding: 0 44px 0 24px;
+		margin-top: 0;
+
+		&__bank,
+		&__usage {
+			padding: 10px 0;
+		}
 
 		&__bank {
 			flex: 1 1 auto;
-			display: flex;
+			border-right: 1px solid $color-secondary-light;
+			display: block;
 			flex-wrap: wrap;
-			padding: 0 20px;
-
-			@media ( min-width: $breakpoint_l ) {
-				display: block;
-				flex-wrap: nowrap;
-			}
+			margin-right: 30px;
 			&-item {
-				white-space: nowrap;
+				float: left;
 				padding-right: 2em;
 
 				&.account .selection-input-text,
@@ -31,7 +32,9 @@
 		}
 
 		&__usage {
-			display: none;
+			flex: 0 0 $form-width;
+			text-align: center;
+			width: $form-width;
 		}
 	}
 
@@ -46,7 +49,7 @@
 			font-size: 12px;
 		}
 		&-input {
-			font-family: $banner-sans-serif;
+			font-family: $banner-main-font-family;
 			position: absolute;
 			top: 0;
 			left: 0;
@@ -76,6 +79,7 @@
 	}
 
 	.application-of-funds-link {
-		display: none;
+		color: $color-black;
+		cursor: pointer;
 	}
 }
