+++ pad/styles_var/SelectGroup.pcss
@@ -1,0 +1,189 @@
+$radio-font-size: 12px;
+
+$radio-border-color: #c8ccd1;
+$radio-background: #f8f9fa;
+
+$radio-border-color-hover: #c8ccd1;
+$radio-background-hover: #eaecf0;
+
+$radio-border-color-active: #386d94;
+$radio-background-active: #f8f9fa;
+
+$radio-border-color-checked: #386d94;
+$radio-background-checked: #5d9fd0;
+$radio-color-checked: #ffffff;
+
+$form-radio-height: 30px;
+
+.wmde-banner {
+
+	.select-group {
+		text-align: center;
+		margin-left: -4px;
+		margin-right: -4px;
+
+		&:after {
+			content: '';
+			display: table;
+			clear: both;
+		}
+
+		&__option {
+			display: block;
+			float: left;
+			line-height: $form-radio-height;
+			padding-left: 4px;
+			padding-right: 4px;
+			padding-bottom: 8px;
+			cursor: pointer;
+			font-size: $radio-font-size;
+			font-weight: bold;
+		}
+
+		&__input {
+			display: none;
+		}
+
+		&__disabled {
+			opacity: 0.3;
+		}
+
+		&__state,
+		&__custom-input {
+			display: block;
+			float: left; /* IE11 fix */
+			width: 100%;
+			border: 1px solid $radio-border-color;
+			border-radius: 2px;
+			background: $radio-background;
+			height: $form-radio-height;
+		}
+
+		&__custom-input {
+			position: relative;
+
+			&--input {
+				display: block;
+				height: $form-radio-height;
+				background: transparent;
+				border: 0;
+				width: 100%;
+				font-size: $radio-font-size;
+				text-align: center;
+				outline: 0;
+				font-weight: bold;
+
+				&::placeholder {
+					opacity: 1;
+					color: $color-black;
+				}
+			}
+
+			&--euro-symbol {
+				position: absolute;
+				top: 0;
+				right: 0;
+				line-height: $form-radio-height;
+				width: 18px;
+				z-index: 10;
+			}
+		}
+
+		.sms-box {
+			padding-bottom: 8px;
+
+			.select-group__option {
+				padding-bottom: 0;
+			}
+
+			a {
+				color: $color-black;
+				&:hover,
+				&:focus {
+					color: $color-black;
+					text-decoration: none;
+				}
+			}
+
+			&__hint {
+				color: $color-gray;
+				font-size: 11px;
+			}
+		}
+
+		&__option:hover .select-group__state,
+		&__option:focus .select-group__state,
+		&__option:hover .select-group__custom-input,
+		&__option:focus .select-group__custom-input {
+			color: $color-black;
+			border-color: $radio-border-color-hover;
+			background: $radio-background-hover;
+		}
+
+		&__option:active .select-group__state,
+		&__custom-input--focused {
+			color: $color-black;
+			border-color: $radio-border-color-active;
+			background: $radio-background-active;
+		}
+
+		&__option:hover .select-group__custom-input--input,
+		&__option:focus .select-group__custom-input--input {
+			color: $color-black;
+		}
+
+		&__option:hover .select-group__custom-input--input::placeholder,
+		&__option:focus .select-group__custom-input--input::placeholder,
+		&__custom-input--focused .select-group__custom-input--input::placeholder {
+			opacity: 0.5;
+		}
+
+		&__custom-input--value-entered .select-group__custom-input--input {
+			color: $radio-color-checked;
+		}
+
+		&__input:checked + .select-group__state,
+		&__custom-input--value-entered {
+			color: $radio-color-checked;
+			border-color: $radio-border-color-checked;
+			background: $radio-background-checked;
+		}
+
+		&__errormessage {
+			display: none;
+			margin-bottom: 8px;
+			font-size: 12px;
+			text-align: center;
+			color: $color-error;
+		}
+
+		&-container--with-error {
+			.select-group__errormessage {
+				display: block;
+			}
+		}
+	}
+
+	.select-group-container--select-amount {
+		margin-top: 0;
+		.select-group__option {
+			width: 25%;
+		}
+
+		.select-group__option--amount-other-input {
+			width: 50%;
+		}
+	}
+
+	.select-group-container--select-interval {
+		.select-group__option {
+			width: 33.333333%;
+		}
+	}
+
+	.select-group-container--select-payment-method {
+		.select-group__option {
+			width: 50%;
+		}
+	}
+}
