+++ pad/styles_var/Slider.pcss
@@ -1,156 +1,122 @@
 .wmde-banner {
-	.keen-slider {
-		display: flex;
-		height: 100%;
-		width: 100%;
-		-webkit-user-select: none;
-		-moz-user-select: none;
-		-ms-user-select: none;
-		user-select: none;
-		-webkit-touch-callout: none;
-		-khtml-user-select: none;
-		touch-action: pan-y;
-		-webkit-tap-highlight-color: transparent;
-		align-items: center;
-	}
-
-	.keen-slider,
-	.keen-slider__slide {
-		overflow: hidden;
-		position: relative;
-		height: 100%;
-	}
-
-	.keen-slider__slide {
-		width: 100%;
-		min-height: 100%;
-		min-width: calc( 100% + 0px );
-		max-width: calc( 100% + 0px );
-	}
-
-	.keen-slider__slide-content {
-		position: relative;
-		height: 100%;
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		cursor: grab;
-		font-family: $banner-serif;
-		font-size: 17px;
-		text-align: center;
-		padding: 36px 30px 30px 36px;
-
-		.headline {
-			font-size: 22px;
-			line-height: 41px;
-
-			@media ( min-width: $breakpoint_xs ) {
-				font-size: 26px;
-			}
-		}
-	}
-
-	.keen-slider[ data-keen-slider-v ] {
-		flex-wrap: wrap;
-	}
-
-	.keen-slider[ data-keen-slider-v ] .keen-slider__slide {
-		width: 100%;
-	}
-
-	.keen-slider[ data-keen-slider-moves ] * {
-		pointer-events: none;
-	}
-
-	.slider-navigation-previous,
-	.slider-navigation-next {
-		position: absolute;
-		z-index: 30;
-		width: 52px;
-		height: 52px;
-		line-height: 52px;
-		top: 50%;
-		margin-top: -26px;
-		border: 0;
-		background: $color-white;
-		border-radius: 50%;
-		text-align: center;
-		padding: 5px 0 0;
-	}
-
-	.slider-navigation-previous {
-		left: -26px;
-		padding-right: 10px;
-		box-shadow: -3px 0 3px rgba( 0, 0, 0, 0.1 );
-		transition: left 500ms $banner-easing;
-
-		&:hover,
-		&:focus {
-			left: -30px;
-		}
-	}
-
-	.slider-navigation-next {
-		right: -26px;
-		padding-left: 10px;
-		box-shadow: 3px 0 3px rgba( 0, 0, 0, 0.1 );
-		transition: right 500ms $banner-easing;
-
-		&:hover,
-		&:focus {
-			right: -30px;
-		}
-	}
-
-	.slideshow-application-of-funds-link {
-		position: absolute;
-		display: inline-block;
-		bottom: 14px;
-		left: 20px;
-		font-size: 12px;
-
-		@media ( min-width: $breakpoint_xs ) {
-			font-size: 15px;
-		}
-	}
-
 	.slider-container {
+		width: 100%;
+		height: 100%;
 		position: relative;
-		padding-bottom: 28px;
-		@media ( min-width: $breakpoint_xs ) {
-			padding-bottom: 38px;
+		font-family: $banner-main-font-family;
+
+		@media ( min-width: $breakpoint_l ) {
+			position: absolute;
+			padding-right: 30px;
+			padding-bottom: 20px;
+		}
+
+		@media ( min-width: $breakpoint_m ) {
+			font-family: $banner-text-font-family;
+			font-size: 1.3em;
+		}
+
+		.navigation-wrapper {
+			display: flex;
+			align-items: center;
+			overflow: hidden;
+			width: 100%;
+			height: 100%;
+			position: relative;
+			padding: 0 0 15px;
+
+			.keen-slider {
+				display: flex;
+				overflow: hidden;
+				position: relative;
+				&__slide {
+					cursor: grab;
+					width: 100%;
+					min-height: 100%;
+					min-width: calc( 100% + 0 );
+					max-width: calc( 100% + 0 );
+					transform: translate3d( 0, 0, 0 );
+					&-content {
+						display: flex;
+						align-items: center;
+						height: 100%;
+						text-align: center;
+						padding: 1em 10px 0;
+						@media ( min-width: $breakpoint_xs ) {
+							padding: 1em 30px 0;
+						}
+						>div {
+							width: 100%;
+						}
+						p {
+							margin: 0 0 1em;
+						}
+					}
+				}
+			}
+
+			.slider-navigation-next,
+			.slider-navigation-previous {
+				position: absolute;
+				display: none;
+				height: 30px;
+				width: 30px;
+				padding: 5px 8px;
+				line-height: 30px;
+				top: 50%;
+				margin-top: -15px;
+
+				@media ( min-width: $breakpoint_xs ) {
+					display: block;
+				}
+			}
+
+			.slider-navigation-previous {
+				left: 0;
+				text-align: left;
+			}
+			.slider-navigation-next {
+				right: 0;
+				text-align: right;
+			}
+		}
+
+		.pagination {
+			position: absolute;
+			bottom: 0;
+			display: flex;
+			justify-content: center;
+			z-index: 999;
+			width: 100%;
+			height: 15px;
+			padding: 0;
+			margin: 0;
+			list-style: none;
+			text-align: center;
+			line-height: 1;
+
+			&-dot {
+				background: $color-gray-light;
+				display: inline-block;
+				width: 15px;
+				height: 15px;
+				margin: 0 4px;
+				border-radius: 50%;
+				cursor: pointer;
+				border: 3px solid $color-white;
+				padding: 0;
+
+				&:focus {
+					outline: 0;
+				}
+			}
+			.is-active {
+				background: $color-white;
+				border: 3px solid $color-primary;
+			}
 		}
 	}
-
-	.pagination {
-		position: absolute;
-		bottom: 0;
-		width: 100%;
-		height: 38px;
-		margin: 0;
-		list-style: none;
-		text-align: center;
-		line-height: 1;
-
-		&-dot {
-			background: $color-white;
-			display: inline-block;
-			width: 12px;
-			height: 12px;
-			padding: 0;
-			margin: 0 4px;
-			border: 2px solid $color-gray;
-			border-radius: 50%;
-			cursor: pointer;
-
-			@media ( min-width: $breakpoint_xs ) {
-				width: 20px;
-				height: 20px;
-			}
-
-			&.is-active {
-				background: $color-gray;
-			}
-		}
+	.headline {
+		font-weight: bold;
 	}
 }
